requirements.txt
# -------------------------
# Core web framework
# -------------------------
fastapi==0.110.2
uvicorn==0.29.0

# -------------------------
# Environment management
# -------------------------
python-dotenv==1.0.1

# -------------------------
# Utilities
# -------------------------
pydantic==2.7.4
faiss-cpu
numpy==1.26.4
openai==1.12.0
pypdf==4.0.1
python-docx==1.1.0
python-multipart==0.0.9
httpx==0.27.2  # pin because openai 1.12 expects 'proxies' arg (httpx 0.28 renamed to 'proxy')
cryptography>=41.0.0

# -------------------------
# Testing
# -------------------------
pytest==8.1.1

# -------------------------
# Auth & persistence
# -------------------------
SQLAlchemy==2.0.29
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.0.0